many: move mount entry utils to osutil
